Fix occlusion culling making entities invisible (read desc)
SE's occlusion culling system does culling for the *next* frame, meaning if it was occluded in the last frame it's also assumed to be occluded this frame, so disoccluded objects have a 1-frame delay in becoming visible. since cameralcd is drawn before the main view, if something is visible to the camera but not the last frame's main view the actor is assumed to be occluded. we patch MyActor.IsOccluded to return false if the current viewid is gbuffer viewid && cameralcd is drawing && occlusion fix is enabled. additionally we skip the occlusion culling pass for camera views since the work is being wasted anyway.